fix: PHPStan (tests/) — appels statiques restants + AuthServiceTest.php
- 37 occurrences supplémentaires de $this->markTestSkipped()/fail()/
  markTestIncomplete() → self:: (10 fichiers) — même motif que les
  assertX() corrigés précédemment, méthodes statiques de
  PHPUnit\Framework\Assert/TestCase, oubliées du premier passage
  (qui ne ciblait que assert*).
- AuthServiceTest.php :
  - 4x assertIsBool()/assertIsString() sans autre assertion dans le
    test → expectNotToPerformAssertions() (le type est déjà garanti
    par la signature de retour, rien de plus à vérifier).
  - 1x assertIsBool() retiré dans un test à 3 sous-cas (les 2 autres
    assertions du test suffisent, celle-ci était redondante).
  - assertTrue(method_exists($this->auth, 'requireAdmin')) retiré :
    la méthode existe garanti statiquement, vérification sans objet.
  - 5x $pdo->query(...)->fetchColumn() + if (!$formId) : chaînage
    direct sur un retour PDOStatement|false, et négation sur un type
    mixte (int|string|false|null) que phpstan-strict-rules interdit.
    Découplé en $stmt puis $stmt === false ? false : fetchColumn(),
    comparaison explicite === false.
- AuditBugsTest.php : cast (int) devenu redondant après le fix
  précédent (PHPStan affine le retour de strtotime() sur une chaîne
  littérale en pur int, plus besoin de caster).

+++ tests/PHPUnit/AuditBugsTest.php
@@ -196,7 +196,7 @@ final class AuditBugsTest extends TestCase
     {
         $tokenId = \generate_uuid();
         $token = \generate_token();
-        $expiresAt = gmdate('Y-m-d H:i:s', (int) strtotime('+30 days'));
+        $expiresAt = gmdate('Y-m-d H:i:s', strtotime('+30 days'));
         $pdo = $this->db->getPdo();
         $pdo->prepare("INSERT INTO tokens (id, submission_id, step_id, email, token, sent_at, expires_at) VALUES (?, ?, ?, ?, ?, datetime('now'), ?)")
             ->execute([$tokenId, $subId, $stepId, $email, $token, $expiresAt]);
